[update login and register pages]
