+++ client/src/components/Ui/Buttons/Buttons.ts
@@ -32,3 +32,21 @@ export const CardEditButton = styled.button`
 export const CardDeleteButton = styled(CardEditButton)`
     background: brown;
 `;
+
+export const LoginButton = styled.button`
+  width: 100%;
+  outline: none;
+  border: none;
+  height: 30px;
+  margin-top: 20px;
+  border-radius: 4px;
+  cursor: pointer;
+
+  color: #fff;
+  background: rgba(0, 0, 0, 0.5);
+  transition: 0.3s;
+
+  &:hover {
+    background: rgba(0, 0, 0, 1);
+  }
+`;
