+++ client/src/components/Ui/Buttons/index.ts
@@ -1,3 +1,13 @@
-import { HeaderButton, CardEditButton, CardDeleteButton } from './Buttons';
+import {
+  HeaderButton,
+  CardEditButton,
+  CardDeleteButton,
+  LoginButton,
+} from './Buttons';
 
-export { HeaderButton, CardEditButton, CardDeleteButton };
+export {
+  HeaderButton,
+  CardEditButton,
+  CardDeleteButton,
+  LoginButton,
+};
