+++ client/src/components/Ui/Forms/Forms.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const LoginForm = styled.form`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  border: 1px solid #000;
+  border-radius: 4px;
+  width: 70%;
+  padding: 20px 30px;
+  margin: 0 auto;
+`;
