+++ client/src/components/Ui/Forms/index.ts
@@ -1,0 +1,3 @@
+import { LoginForm } from './Forms';
+
+export { LoginForm };
