+++ client/src/components/Ui/Inputs/index.ts
@@ -1,0 +1,3 @@
+import { InputStyled } from './Inputs';
+
+export { InputStyled };
