+++ client/src/components/Ui/Typography/Typography.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const LoginTitle = styled.h1`
+  font-size: 1.5rem;
+  color: #000;
+  margin-bottom: 20px;
+  text-align: center;
+`;
