+++ client/src/components/Ui/Typography/index.ts
@@ -1,0 +1,3 @@
+import { LoginTitle } from './Typography';
+
+export { LoginTitle };
