+++ client/src/pages/Auth/Login/styles.ts
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import { Input } from '@material-ui/core';
 
 export const LoginPage = styled.div`
   position: fixed;
@@ -10,7 +9,6 @@ export const LoginPage = styled.div`
   left: 0;
   right: 0;
   bottom: 0;
-  background: rgba(0,0,0, 0.8);
 
   display: flex;
   justify-content: center;
@@ -23,50 +21,6 @@ export const Wrapper = styled.div`
   width: 100%;
   display: flex;
   flex-direction: column;
-`;
-
-export const Title = styled.h1`
-  color: #fff;
-  margin-bottom: 20px;
-  text-align: center;
-`;
-
-export const LoginForm = styled.form`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 50%;
-  background: #fff;
-  padding: 20px 30px;
-  margin: 0 auto;
-`;
-
-export const InputStyled = styled(Input)`
-  width: 100%;
-  height: 30px;
-  margin: 10px 0;
-
-  &::placeholder {
-    padding: 0 5px;
-  }
-`;
-
-export const LoginButton = styled.button`
-  width: 100%;
-  outline: none;
-  border: none;
-  height: 30px;
-  margin-top: 20px;
-  border-radius: 4px;
-  cursor: pointer;
-
-  color: #fff;
-  background: rgba(0, 0, 0, 0.5);
-  transition: 0.3s;
-
-  &:hover {
-    background: rgba(0, 0, 0, 1);
-  }
 `;
 
 export const RegisterLink = styled(Link)`
