+++ client/src/pages/Auth/Register/styles.ts
@@ -9,7 +9,6 @@ export const LoginPage = styled.div`
   left: 0;
   right: 0;
   bottom: 0;
-  background: rgba(0,0,0, 0.8);
 
   display: flex;
   justify-content: center;
